requirement.txt
Django==1.9.2
django-rest-swagger==0.3.5
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
rest-social-auth==0.4.1
